Struct mysql_async::binlog::events::ExecuteLoadQueryEvent

pub struct ExecuteLoadQueryEvent<'a> { /* private fields */ }

Execute load query event.

Used for LOAD DATA INFILE statements as of MySQL 5.0.

It similar to Query_log_event but before executing the query it substitutes original filename in LOAD DATA query with name of temporary file.

Implementations

impl<'a> ExecuteLoadQueryEvent<'a>

pub fn new(
file_id: u32,
dup_handling: LoadDuplicateHandling,
status_vars: impl Into<Cow<'a, [u8]>>,
schema: impl Into<Cow<'a, [u8]>>
) -> ExecuteLoadQueryEvent<'a>

Creates a new instance.

pub fn with_thread_id(self, thread_id: u32) -> ExecuteLoadQueryEvent<'a>

Sets the thread_id value.

pub fn with_execution_time(
self,
execution_time: u32
) -> ExecuteLoadQueryEvent<'a>

Sets the execution_time value.

pub fn with_error_code(self, error_code: u16) -> ExecuteLoadQueryEvent<'a>

Sets the error_code value.

pub fn with_file_id(self, file_id: u32) -> ExecuteLoadQueryEvent<'a>

Sets the file_id value.

pub fn with_start_pos(self, start_pos: u32) -> ExecuteLoadQueryEvent<'a>

Sets the start_pos value.

pub fn with_end_pos(self, end_pos: u32) -> ExecuteLoadQueryEvent<'a>

Sets the end_pos value.

pub fn with_dup_handling(
self,
dup_handling: LoadDuplicateHandling
) -> ExecuteLoadQueryEvent<'a>

Sets the dup_handling value.

pub fn with_status_vars(
self,
status_vars: impl Into<Cow<'a, [u8]>>
) -> ExecuteLoadQueryEvent<'a>

Sets the status_vars value (max length is `u16::MAX).

pub fn with_schema(
self,
schema: impl Into<Cow<'a, [u8]>>
) -> ExecuteLoadQueryEvent<'a>

Sets the schema value (max length is `u8::MAX).

pub fn with_query(
self,
query: impl Into<Cow<'a, [u8]>>
) -> ExecuteLoadQueryEvent<'a>

Sets the query value.

pub fn thread_id(&self) -> u32

Returns the thread_id value.

thread_id is the ID of the thread that issued this statement. It is needed for temporary tables.

pub fn execution_time(&self) -> u32

Returns the execution_time value.

execution_time is the time from when the query started to when it was logged in the binlog, in seconds.

pub fn error_code(&self) -> u16

Returns the error_code value.

error_code is the error code generated by the master. If the master fails, the slave will fail with the same error code.

pub fn file_id(&self) -> u32

Returns the file_id value.

file_id is the ID of the temporary file to load.

pub fn start_pos(&self) -> u32

Returns the start_pos value.

start_pos is the start position within the statement for filename substitution.

pub fn end_pos(&self) -> u32

Returns the end_pos value.

end_pos is the end position within the statement for filename substitution.

pub fn dup_handling(&self) -> LoadDuplicateHandling

Returns the dup_handling value.

dup_handling represents the information on how to handle duplicates.

pub fn status_vars_raw(&'a self) -> &'a [u8]

Returns the status_vars value.

status_vars contains zero or more status variables. Each status variable consists of one byte identifying the variable stored, followed by the value of the variable.

pub fn status_vars(&'a self) -> &'a StatusVars<'a>

Returns an iterator over status variables.

pub fn schema_raw(&'a self) -> &'a [u8]

Returns the schema value.

schema is schema name.

pub fn schema(&'a self) -> Cow<'a, str>

Returns the schema value as a string (lossy converted).

pub fn query_raw(&'a self) -> &'a [u8]

Returns the query value.

query is the corresponding LOAD DATA INFILE statement.

pub fn query(&'a self) -> Cow<'a, str>

Returns the query value as a string (lossy converted).

pub fn into_owned(self) -> ExecuteLoadQueryEvent<'static>