Module mysql_async::consts

Structs

CapabilityFlags

Client capability flags

ColumnFlags

MySql column flags

CursorType

Mysql cursor type.

Flags2

Bitmask of flags that are usually set with SET.

SqlMode

Bitmask of flags that are usually set with SET sql_mode.

StatusFlags

MySql server status flags

StmtExecuteParamFlags

MySql stmt execute params flags.

StmtExecuteParamsFlags

MySql stmt execute params flags.

UnknownCapabilityFlags

UnknownColumnFlags

UnknownColumnType

UnknownCursorType

UnknownFlags2

UnknownGeometryType

UnknownItemResultType

UnknownSessionStateType

UnknownSqlMode

UnknownStatusFlags

UnknownStmtExecuteParamFlags

UnknownStmtExecuteParamsFlags

Enums

ColumnType

Type of MySql column field

Command

MySql server commands

GeometryType

Geometry type.

ItemResult

Type of the user defined function return slot and arguments.

SessionStateType

Type of state change information (part of MySql’s Ok packet).

Statics

DEFAULT_MAX_ALLOWED_PACKET

MAX_PAYLOAD_LEN

MIN_COMPRESS_LENGTH

UTF8MB4_GENERAL_CI

UTF8_GENERAL_CI