Module mysql_async::prelude

Traits used in this crate

Modules

params

Macros

params
This macro is a convenient way to pass named parameters to a statement.

Traits

ConnectionLike
Everything that is connection.

ConvIr
Basic operations on FromValue conversion intermediate result.

FromRow
Trait to convert Row into a tuple of FromValue implementors up to arity 12.

FromValue
Implement this trait to convert value to something.

LocalInfileHandler
Trait used to handle local infile requests.

Protocol
Trait for protocol markers crate::TextProtocol and crate::BinaryProtocol.

Queryable
Represents something queryable like connection or transaction.

ToValue