mysql_async
Struct Column

pub struct Column { /* private fields */ }

Represents MySql Column (column packet).

Implementations

impl Column

pub fn new(column_type: ColumnType) -> Column

pub fn with_schema(self, schema: &[u8]) -> Column

pub fn with_table(self, table: &[u8]) -> Column

pub fn with_org_table(self, org_table: &[u8]) -> Column

pub fn with_name(self, name: &[u8]) -> Column

pub fn with_org_name(self, org_name: &[u8]) -> Column

pub fn with_flags(self, flags: ColumnFlags) -> Column

pub fn with_column_length(self, column_length: u32) -> Column

pub fn with_character_set(self, character_set: u16) -> Column

pub fn with_decimals(self, decimals: u8) -> Column

pub fn column_length(&self) -> u32

Returns value of the column_length field of a column packet.

Can be used for text-output formatting.

pub fn column_type(&self) -> ColumnType

Returns value of the column_type field of a column packet.

pub fn character_set(&self) -> u16

Returns value of the character_set field of a column packet.

pub fn flags(&self) -> ColumnFlags

Returns value of the flags field of a column packet.

pub fn decimals(&self) -> u8

Returns value of the decimals field of a column packet.

Max shown decimal digits. Can be used for text-output formatting

0x00 for integers and static strings

0x1f for dynamic strings, double, float

0x00..=0x51 for decimals

pub fn schema_ref(&self) -> &[u8]

Returns value of the schema field of a column packet as a byte slice.

pub fn schema_str(&self) -> Cow<'_, str>

Returns value of the schema field of a column packet as a string (lossy converted).

pub fn table_ref(&self) -> &[u8]

Returns value of the table field of a column packet as a byte slice.

pub fn table_str(&self) -> Cow<'_, str>

Returns value of the table field of a column packet as a string (lossy converted).

pub fn org_table_ref(&self) -> &[u8]

Returns value of the org_table field of a column packet as a byte slice.

โorg_tableโ is for original table name.

pub fn org_table_str(&self) -> Cow<'_, str>

Returns value of the org_table field of a column packet as a string (lossy converted).

pub fn name_ref(&self) -> &[u8]

Returns value of the name field of a column packet as a byte slice.

pub fn name_str(&self) -> Cow<'_, str>

Returns value of the name field of a column packet as a string (lossy converted).

pub fn org_name_ref(&self) -> &[u8]

Returns value of the org_name field of a column packet as a byte slice.

โorg_nameโ is for original column name.

pub fn org_name_str(&self) -> Cow<'_, str>

Returns value of the org_name field of a column packet as a string (lossy converted).