Struct mysql_async::Transaction

pub struct Transaction<'a>(_);

This struct represents MySql transaction.

Transaction is just a sugar for START TRANSACTION, ROLLBACK and COMMIT queries, so please note, that it is easy to mess things up calling this queries manually.

You should alwasy call either commit or rollback, otherwise transaction will be rolled back implicitly when corresponding connection is dropped or queried.

Implementations

impl<'a> Transaction<'a>

pub async fn commit(self) -> Result<()>

Performs COMMIT query.

pub async fn rollback(self) -> Result<()>

Performs ROLLBACK query.