mysql_async::binlog
Module events

Structs

AnonymousGtidEvent

Anonymous GTID event.

BeginLoadQueryEvent

Begin load query event.

BinlogEventFooter

Binlog event footer.

BinlogEventHeader

The binlog event header starts each event and is 19 bytes long assuming binlog version >= 4.

ColumnCharsets

Contains charset+collation for column.

ColumnName

Name of a column in ColumnNames.

ColumnNames

Contains names of columns.

DefaultCharset

Optional metadata field that contains charsets for columns.

DeleteRowsEvent

Delete rows event.

DeleteRowsEventV1

Delete rows event v1 (mariadb and mysql 5.1.15-5.6.x).

EnumStrValue

String value for an ENUM column variant.

EnumStrValues

Contains string values for ENUM column.

EnumsStrValues

Contains string values of ENUM columns.

Event

Raw binlog event.

ExecuteLoadQueryEvent

Execute load query event.

FormatDescriptionEvent

A format description event is the first event of a binlog for binlog-version 4.

GeometryTypes

Contains real types for every geometry column.

GtidEvent

GTID stands for Global Transaction IDentifier.

IncidentEvent

Used to log an out of the ordinary event that occurred on the master.

IntvarEvent

Integer based session-variables event.

IterCharsets

IterEnumStrValues

IterGeometryTypes

IterIndexes

IterKeys

IterNames

IterNonDefault

IterSetStrValues

NonDefaultCharset

Contains column_id -> charset+collation mapping for columns with non-default charsets. (see DefaultCharset::iter_non_default).

OptionalMetaExtractor

Helper struct that extracts optional metadata for columns.

OptionalMetadataIter

Iterator over fields of an optional metadata.

PartialUpdateRowsEvent

Partial update rows event.

PrimaryKeyWithPrefix

Info about primary key with a prefix (see PrimaryKeysWithPrefix).

PrimaryKeysWithPrefix

Contains a sequence of primary keys with prefix.

QueryEvent

A query event is created for each query that modifies the database, unless the query is logged row-based.

RandEvent

Rand event.

RotateEvent

The rotate event is added to the binlog as last event to tell the reader what binlog to request next.

RowsEvent

Common base structure for all row-containing binary log events.

RowsEventRows

Iterator over rows of a RowsEvent.

RowsQueryEvent

Query that caused the following ROWS_EVENT.

SetStrValue

String value for a SET column variant.

SetStrValues

Contains string values for SET column.

SetsStrValues

Contains string values of SET columns.

SimplePrimaryKey

Contains a sequence of PK column indexes where PK doesn’t have a prefix.

StatusVar

Raw status variable.

StatusVars

Status variables of a QueryEvent.

StatusVarsIterator

Iterator over status vars of a QueryEvent.

TableMapEvent

Table map event.

TransactionPayloadEvent

Event that encloses all the events of a transaction.

TransactionPayloadReader

This structure implements io::BufRead and represents the payload of a TransactionPayloadEvent.

UpdateRowsEvent

Update rows event.

UpdateRowsEventV1

Update rows event v1 (mariadb and mysql 5.1.15-5.6.x).

UserVarEvent

User variable event.

WriteRowsEvent

Write rows event.

WriteRowsEventV1

Write rows event v1 (mariadb and mysql 5.1.15-5.6.x).

XidEvent

Xid event.

Enums

BadColumnType

EventData

Parsed event data.

OptionalMetadataField

RowsEventData

Rows events are unified under this enum (see EventData).

StatusVarVal

Status variable value.