mysql_async
Module prelude

Traits used in this crate

Modules

params

Macros

params

This macro is a convenient way to pass named parameters to a statement.

Traits

AsQuery

Types that can be treated as a MySQL query.

BatchQuery

Helper trait for batch statement execution.

FromRow

Trait to convert Row into a tuple of FromValue implementors up to arity 12.

FromValue

Implement this trait to convert a value to some type.

GlobalHandler

Global, Opts-level LOCAL INFILE handler (see “LOCAL INFILE Handlers” section of the README.md).

Protocol

Trait for protocol markers crate::TextProtocol and crate::BinaryProtocol.

Query

MySql text query.

Queryable

Methods of this trait are used to execute database queries.

StatementLike

Everything that is a statement.

ToConnection

Everything that is a connection.

ToValue

WithParams

Helper, that constructs QueryWithParams.

Derive Macros

FromRow

Derives FromRow. See mysql_common crate-level docs for more info.

FromValue

Derives FromValue. See mysql_common crate-level docs for more info.