Module mysql_async::prelude

Traits used in this crate

Modules

params

Macros

params

This macro is a convenient way to pass named parameters to a statement.

Traits

BatchQuery

Helper trait for batch statement execution.

ConvIr

Basic operations on FromValue conversion intermediate result.

FromRow

Trait to convert Row into a tuple of FromValue implementors up to arity 12.

FromValue

Implement this trait to convert value to something.

LocalInfileHandler

Trait used to handle local infile requests.

Protocol

Trait for protocol markers crate::TextProtocol and crate::BinaryProtocol.

Query

MySql text query.

Queryable

Methods of this trait are used to execute database queries.

StatementLike

Everything that is a statement.

ToConnection

Everything that is a connection.

ToValue

WithParams

Helper, that constructs QueryWithParams.