Struct mysql_async::Transaction

pub struct Transaction<T>(_);

This struct represents MySql transaction.

Transaction it's a sugar for START TRANSACTION, ROLLBACK and COMMIT queries, so one should note that it is easy to mess things up calling this queries manually. Also you will get NestedTransaction error if you call transaction.start_transaction(_).

Methods

impl<T: Queryable + ConnectionLike> Transaction<T>

pub async fn commit(self) -> Result<T>

Returns future that will perform COMMIT query and resolve to a wrapped Queryable.

pub async fn rollback(self) -> Result<T>

Returns future that will perform ROLLBACK query and resolve to a wrapped Queryable.